Split and format reviewed cleanup tests
Preserve the 22 focused cleanup tests while separating contract, execution, CLI, shared support, and workflow-policy coverage into formatter-compliant modules.
